PHP 7 Memcached: fix calling `getMulti` with too many params
At the time of this writing, the PHP 7 branch of the memcached client
has a slightly different API than `master`. `getMulti()` takes 2
arguments instead of 3 (it's missing the second `$cas_tokens` argument).

We want Matryoshka to work with both PHP 7 and PHP 5 clients, and this
is the cleanest solution we could think of.