Add delete before update backend wrapper
Does an add delete add on each of the keys in the case that someone is
trying to update an existing key. When switching to mcrouter, we only
have a reliable delete stream, we don't have a way to see if any other
requests to a memcache machine have failed, so this wrapper will ensure
that when updating something, we're always doing a delete and add
instead of updating a key.