Reject McRouter invalid '\r' in
Inject an extended list of invalid characters to KeyFix in the McRouter
factory. McRouter doesn't seem to want to play nice with the "\r"
character. It's the only character that was failing in tests.

This now passes all tests with Memcached _and_ McRouter as the backend
service.

We need to override the factory method to pass the `invalidCharacters`
parameter to the KeyFix wrapper constructor. This is accomplished inline
around the constructor call. That means breaking out the constructor
call back to the class factory method.

I tried to avoid this, but late static binding was trying to
call the parent class inside the call to the KeyFix constructor.