Scope: make key more random
I think it is actually possible there's cache collisions happening here.
We were generating a cache scope which is only an 8 character hex value.

If the number of things we're caching under one cache scope is in the
hundreds of thousands or millions I think there could be multiple collisions:
https://en.wikipedia.org/wiki/Birthday_problem#Probability_table