Scope: minimize the number of gets
If getting the scope prefix value is a miss, then we all keys under the
scope are by definition a miss, so we don't need to bother looking.

This defers the generating and storing of the scope prefix value till
we actually need it (on a set). In practice, this will reduce the number
of gets if using getMultiple() or if you don't always *set()* on a miss.